+++ doc/qbk/main.qbk
@@ -131,6 +131,7 @@
 
 [import ../../test/doc/core_1_refresher.cpp]
 [import ../../test/doc/core_3_layers.cpp]
+[import ../../test/doc/http_10_custom_parser.cpp]
 [import ../../test/doc/websocket_3_handshake.cpp]
 
 [import ../../include/boost/beast/core/detect_ssl.hpp]
