+++ example/doc/http_examples.hpp
@@ -860,199 +860,6 @@ do_form_request(
 
 //]
 
-
-
-//------------------------------------------------------------------------------
-//
-// Example: Custom Parser
-//
-//------------------------------------------------------------------------------
-
-//[example_http_custom_parser
-
-template<bool isRequest>
-class custom_parser : public basic_parser<isRequest>
-{
-private:
-    /// Called after receiving the request-line (isRequest == true).
-    void
-    on_request_impl(
-        verb method,                // The method verb, verb::unknown if no match
-        string_view method_str,     // The method as a string
-        string_view target,         // The request-target
-        int version,                // The HTTP-version
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called after receiving the start-line (isRequest == false).
-    void
-    on_response_impl(
-        int code,                   // The status-code
-        string_view reason,         // The obsolete reason-phrase
-        int version,                // The HTTP-version
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called after receiving a header field.
-    void
-    on_field_impl(
-        field f,                    // The known-field enumeration constant
-        string_view name,           // The field name string.
-        string_view value,          // The field value
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called after the complete header is received.
-    void
-    on_header_impl(
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called just before processing the body, if a body exists.
-    void
-    on_body_init_impl(
-        boost::optional<
-            std::uint64_t> const&
-                content_length,     // Content length if known, else `boost::none`
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called for each piece of the body, if a body exists.
-    //!
-    //! This is used when there is no chunked transfer coding.
-    //!
-    //! The function returns the number of bytes consumed from the
-    //! input buffer. Any input octets not consumed will be will be
-    //! presented on subsequent calls.
-    //!
-    std::size_t
-    on_body_impl(
-        string_view s,              // A portion of the body
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called for each chunk header.
-    void
-    on_chunk_header_impl(
-        std::uint64_t size,         // The size of the upcoming chunk,
-                                    // or zero for the last chunk
-        string_view extension,      // The chunk extensions (may be empty)
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called to deliver the chunk body.
-    //!
-    //! This is used when there is a chunked transfer coding. The
-    //! implementation will automatically remove the encoding before
-    //! calling this function.
-    //!
-    //! The function returns the number of bytes consumed from the
-    //! input buffer. Any input octets not consumed will be will be
-    //! presented on subsequent calls.
-    //!
-    std::size_t
-    on_chunk_body_impl(
-        std::uint64_t remain,       // The number of bytes remaining in the chunk,
-                                    // including what is being passed here.
-                                    // or zero for the last chunk
-        string_view body,           // The next piece of the chunk body
-        error_code& ec) override;   // The error returned to the caller, if any
-
-    /// Called when the complete message is parsed.
-    void
-    on_finish_impl(
-        error_code& ec) override;   // The error returned to the caller, if any
-
-public:
-    custom_parser() = default;
-};
-
-//]
-
-// Definitions are not part of the docs but necessary to link
-
-template<bool isRequest>
-void custom_parser<isRequest>::
-on_request_impl(verb method, string_view method_str,
-    string_view path, int version, error_code& ec)
-{
-    boost::ignore_unused(method, method_str, path, version);
-    ec = {};
-}
-
-template<bool isRequest>
-void custom_parser<isRequest>::
-on_response_impl(
-    int status,
-    string_view reason,
-    int version,
-    error_code& ec)
-{
-    boost::ignore_unused(status, reason, version);
-    ec = {};
-}
-
-template<bool isRequest>
-void custom_parser<isRequest>::
-on_field_impl(
-    field f,
-    string_view name,
-    string_view value,
-    error_code& ec)
-{
-    boost::ignore_unused(f, name, value);
-    ec = {};
-}
-
-template<bool isRequest>
-void custom_parser<isRequest>::
-on_header_impl(error_code& ec)
-{
-    ec = {};
-}
-
-template<bool isRequest>
-void custom_parser<isRequest>::
-on_body_init_impl(
-    boost::optional<std::uint64_t> const& content_length,
-    error_code& ec)
-{
-    boost::ignore_unused(content_length);
-    ec = {};
-}
-
-template<bool isRequest>
-std::size_t custom_parser<isRequest>::
-on_body_impl(string_view body, error_code& ec)
-{
-    boost::ignore_unused(body);
-    ec = {};
-    return body.size();
-}
-
-template<bool isRequest>
-void custom_parser<isRequest>::
-on_chunk_header_impl(
-    std::uint64_t size,
-    string_view extension,
-    error_code& ec)
-{
-    boost::ignore_unused(size, extension);
-    ec = {};
-}
-
-template<bool isRequest>
-std::size_t custom_parser<isRequest>::
-on_chunk_body_impl(
-    std::uint64_t remain,
-    string_view body,
-    error_code& ec)
-{
-    boost::ignore_unused(remain);
-    ec = {};
-    return body.size();
-}
-
-template<bool isRequest>
-void custom_parser<isRequest>::
-on_finish_impl(error_code& ec)
-{
-    ec = {};
-}
-
 //------------------------------------------------------------------------------
 //
 // Example: Incremental Read
