+++ test/doc/Jamfile
@@ -21,6 +21,7 @@ alias run-tests :
     [ compile websocket_snippets.cpp ]
     [ run core_1_refresher.cpp $(TEST_MAIN) ]
     [ run core_3_layers.cpp $(TEST_MAIN) ]
+    [ run http_10_custom_parser.cpp $(TEST_MAIN) ]
     [ run http_examples.cpp $(TEST_MAIN) ]
     [ run websocket_3_handshake.cpp $(TEST_MAIN) ]
     ;
@@ -29,6 +30,7 @@ exe fat-tests :
     $(TEST_MAIN)
     core_1_refresher.cpp
     core_3_layers.cpp
+    http_10_custom_parser.cpp
     http_examples.cpp
     websocket_3_handshake.cpp
     ;
